[BUGFIX] Properly apply mount points in RecordLinkHandler
The RecordLinkHandler allows to define custom mount
points, to be used for the page tree in the LinkBrowser.

However, due to not properly handling an empty
Configuration, the page tree was no longer shown
in case no mount points were defined, because
this led to the value being "[0]".

This is now fixed by properly handling an empty
mount point configuration.

Resolves: #96190
Releases: main, 11.5
Change-Id: Ic1a01f11c7499262697110664a5eb6cea94d13e1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72450
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>