[BUGFIX] Replace undefined variable with reading PageTSConfig option
Long story short, through multiple refactorings along the road, from
the inital introduction of the 'mod.newContentElementWizard.override'
option with #70055 over streamlining the TSConfig api (#85016), a lot
of stations in between which already broke reading that option, which
was used for adding the 'new record button', this was refactored into
a dedicated 'createAddButtonForTable()' method with #94456, ending up
with even more broken code and leaving an undefined variable $tsConfig.

Instead of removing the unused variable we load the pageTs here to be
able to read this option again and restore the intended behaviour.

Resolves: #96596
Related: #94456
Related: #85016
Related: #70055
Releases: main, 11.5
Change-Id: I91b24bf338c6531ea220a41dca2076b71a7da96f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73096
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>