[TASK] Replace 'CompositeExpression->add()' usage
doctrine/dbal moves their classes forward and deprecated
some stuff while providing replacement for it. This has
been early adopted for the core facade classes with #97081
in a forward-compatible way. Before deprecating the old
methods core usage have to be replaced.

This patch replaces the usage of 'CompositeExpression::add()'
with the immutable way using 'with()' with re-assigning the
returned instance to the used variable.

In general this means that following usage example:

  $or = CompositeExpression::or();
  $or->add($queryBuilder->expr()->eq(...));

will be replaced with:

  $or = CompositeExpression::or();
  $or = $or->with($queryBuilder->expr()->eq(...));

Alternativly this could be refactored collecting parts in a
array first and add them later with one 'with()' call or by
directly creating the 'or()' or 'and()' CompositeExpression
with the provided parts directly - which would have been a
larger and harder to review patch. Thus this patch does the
simple replacement to have minimal changes for easier review
experience.

Resolves: #97124
Related: #97081
Releases: main
Change-Id: I00fde09593e1c1ccb2949daaec3ee4aa688dda63
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73833
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>