[FEATURE] Introduce TCA type "link"
In our process of using dedicated TCA types,
the new TCA type "link" is introduced and
replaces "renderType=link" of TCA type "input".

Notable changes to the previous "inputLink" behaviour:

- "linkPopup" fieldControl is now always used by the
  FormEngine element
- Corresponding fieldControl options are migrated
  to columns config
- The type now makes use of include lists instead of
  exclude lists
- The allowed link types are now evaluated in DataHandler
- "trim" is always done in DataHandler and FormEngine
- "max" is no longer evaluated
- "softref=typolink" is automatically set

As mentioned, the TCA type "link" uses include lists.
This also required to adjust the LinkPopup fieldControl,
as well as the Link Browser classes. The previously
used "blinkLink[Fields|Options]" options are still
supported in case the "LinkPopup" fieldControl is used
for another TCA type. However, they might be deprecated
soon.

Finally, the TSconfig key for the MailLinkHandler is
renamed from "mail" to "email", to unify the naming
in the codebase.

Resolves: #97159
Releases: main
Change-Id: Ied9e957b973ed8904736e5d4353a989ea76065d8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73866
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Oliver Bartsch <bo@cedev.de>