[!!!][TASK] Migrate backend modal component to lit
Avoid jQuery and especially the bootstrap
jQuery compat layer.
Use lit as template engine instead of jQuery composition
and avoid the bootstrap jQuery compat layer for initialization.

Will allow lit's TemplateResult to be directly passed
to the modal as content, allowing for a secure and
straight forward templating system for the modal API.

Left for another patch is the migration of
AbstractInteractableModule away from jQuery
and the adaption of login-refresh.ts to use
the TYPO3 Modal api, instead of directly calling
into bootstraps jQuery-Modal API.

Support for jQuery based modal content is kept for
now, as it doesn't require us to load jQuery
proactively and may ease migration for some users,
but lit TemplateResults are suggested to be used
instead. The jQuery content support may be
deprecated in an upcoming sprint release.

Releases: main
Resolves: #98288
Change-Id: Ic212485293b4cbe86f8a855d40de4c72f8de9558
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74165
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Jonas Eberle <flightvision@googlemail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>