[BUGFIX] Use correct table identifier for collapse state in record list
The "table" for translated page records is internally identified as
"pages_translated". However, the real table name "pages" was used to
check the collapse state of the table in the record list module.

The correct table identifier is now used to check if a table is
collapsed or expanded.

Resolves: #98634
Releases: main, 11.5
Change-Id: Iefb3e56228e34759cbddc92d769074579de36f85
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76152
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>